refactor: Input 컴포넌트를 presentational로 분리한다
